\name{jpstat.default}
\alias{jpstat.default}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
make jpstat object
}
\description{
make jpstat object
}
\usage{
jpstat.default(data, id, stat.name, gov, statistics.name, title, survey.date)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
list of dataframe
}
  \item{id}{
data ID
}
  \item{stat.name}{
Stat name
}
  \item{gov}{
government 
}
  \item{statistics.name}{
statistics name
}
  \item{title}{
title
}
  \item{survey.date}{
survey date
}
}
\details{
This function is only for internal usages.
}
\value{
jpstat object
}
\author{
Kiwamu Ishikura <ishikura.kiwamu@gmail.com>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~jpstat }
