\name{jpstat}
\alias{jpstat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Make jpstat object
}
\description{
Make jpstat object
}
\usage{
jpstat(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
required arguments
}
}
\details{
This function is only for internal usages.
}
\value{
jpstat object
}
\author{
Kiwamu Ishikura <ishikura.kiwamu@gmail.com>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~jpstat }
