\name{print.jpstat}
\docType{methods}
\alias{print.jpstat}
\alias{print,jpstat-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print object of jpstat
}
\description{
print object of jpstat
}
\usage{
\S4method{print.jpstat}(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
object of jpstat
}
}
\value{
  print
}
\author{
Kiwamu Ishikura <ishikura.kiwamu@gmail.com>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{getStatsData}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~getStatsData }
