\name{print.jpstat}
\alias{print.jpstat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print class of jpstat
}
\description{
print class of jpstat
}
\usage{
\method{print}{jpstat}(jpstat, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{jpstat}{
object of jpstat
}
  \item{\dots}{
aditional arguments
}
}
\value{
  print
}
\author{
Kiwamu Ishikura <ishikura.kiwamu@gmail.com>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{getStatsData}
}
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~getStatsData }
