\name{print.jpstat}
\alias{print.jpstat}
\encoding{UTF-8}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
print class of jpstat
}
\description{
print class of jpstat
}
\usage{
print.jpstat(jpstat, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{jpstat}{
object of jpstat class
}
}
\value{
  print
}
\references{
  http://statdb.nstac.go.jp/system-info/api/api-spec/
}
\author{
Kiwamu Ishikura <ishikura.kiwamu@gmail.com>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~getStatsData }
